Tweaked GameVar, TimeUtils, PPlayerLevels
Added Generic GetValue functions in GameVar.
Modified FromUnixTime function to allow different values.

+++ Assets/Plugins/Playtomic/PlayerAward.cs
@@ -12,7 +12,7 @@ public class PlayerAward : PDictionary
 		{			
 			if(x == "date") 
 			{
-			    date = TimeUtils.FromUnixTime((int)data[x]);
+			    date = TimeUtils.FromUnixTime(data[x]);
 				continue;
 			}
